Remove configuration test (see commit comment)
The configuration class can not be compiled by itself- it depends on
the Service class to define the `logdna-agent` service so that when the
configuration changes it can restart the service.

This class *is* still tested, as it is called by the `logdna` class and
thus falls under that test.